chore: move bridge types to tangle-shared-ui types
