+++ apps/tangle-dapp/src/pages/bridge/hooks/useHyperlaneQuote.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 import { getHyperlaneWarpCore } from '../lib/hyperlane/context';
-import { BridgeTokenType } from '../types';
+import { BridgeTokenType } from '@webb-tools/tangle-shared-ui/types';
 import { getHyperlaneChainName, tryFindToken } from '../lib/hyperlane/utils';
 
 export type HyperlaneQuoteProps = {
