+++ libs/tangle-shared-ui/src/hooks/useLocalStorage.ts
@@ -4,7 +4,7 @@ import { HexString } from '@polkadot/util/types';
 import { useCallback, useEffect, useState } from 'react';
 
 import Optional from '../utils/Optional';
-import { Payout, TangleTokenSymbol } from '../types';
+import { BalanceType, Payout, TangleTokenSymbol } from '../types';
 import {
   Collator,
   Dapp,
@@ -83,7 +83,7 @@ export type LocalStorageValueOf<T extends LocalStorageKey> =
                         : T extends LocalStorageKey.ONBOARDING_MODALS_SEEN
                           ? OnboardingPageKey[]
                           : T extends LocalStorageKey.EVM_TOKEN_BALANCES
-                            ? Record<string, any>
+                            ? Record<string, BalanceType>
                             : never;
 
 export const getJsonFromLocalStorage = <Key extends LocalStorageKey>(
