+++ libs/tangle-shared-ui/src/types/index.ts
@@ -1,7 +1,14 @@
 import type { BN } from '@polkadot/util';
 import type { WebbProviderType } from '@webb-tools/abstract-api-provider/types';
 import { SubstrateAddress } from '@webb-tools/webb-ui-components/types/address';
-import { EVMTokenBridgeEnum } from '@webb-tools/evm-contract-metadata';
+
+import { PresetTypedChainId } from '@webb-tools/dapp-types';
+import {
+  EVMTokenBridgeEnum,
+  EVMTokenEnum,
+} from '@webb-tools/evm-contract-metadata';
+import { Abi } from 'viem';
+import { Decimal } from 'decimal.js';
 
 export type TangleTokenSymbol = 'tTNT' | 'TNT';
 
@@ -24,6 +31,8 @@ export const ExplorerType = {
   Substrate: 'polkadot' as WebbProviderType,
   EVM: 'web3' as WebbProviderType,
 } as const;
+
+// Bridge
 
 export enum BridgeTxState {
   Initializing = 'Initializing',
@@ -50,3 +59,32 @@ export type BridgeQueueTxItem = {
   destinationTxExplorerUrl?: string;
   bridgeType: EVMTokenBridgeEnum;
 };
+
+export interface BridgeTokenType {
+  tokenSymbol: string;
+  tokenType: EVMTokenEnum;
+  bridgeType: EVMTokenBridgeEnum;
+  address: `0x${string}`;
+  abi: Abi;
+  decimals: number;
+  chainId: PresetTypedChainId;
+  hyperlaneRouteContractAddress?: `0x${string}`;
+}
+
+export type BridgeChainsConfigType = Record<
+  PresetTypedChainId,
+  Record<
+    PresetTypedChainId,
+    {
+      supportedTokens: BridgeTokenType[];
+    }
+  >
+>;
+
+export type TokenBalanceType = BridgeTokenType & {
+  balance: Decimal;
+};
+
+export type BalanceType = Partial<
+  Record<PresetTypedChainId, TokenBalanceType[]>
+>;
